Implemented LSTM and GRU models for stock prediction and created a template for ProphetModel class. Started working on ARIMA model.
I reviewed the code for the BasePredictor, GRUPredictor, and LSTMPredictor classes, including their respective train and optimize_model methods.

I provided an example of how to use the GRUPredictor and LSTMPredictor classes to train and evaluate stock prediction models using time series data.

I started working on the ARIMA model. Although it is not fully functional yet, I have addressed a TypeError by updating the ARIMAModel's init() method to accept the correct number of parameters (p, d, and q).

I implemented a template for the ProphetModel class, which serves as a starting point for creating a stock prediction model using the Facebook Prophet library. This class is a subclass of the Model class and includes placeholder methods for training and prediction.