Modularized prediction code and added worked on visualizer
The code is now more modular, and it will be easier to maintain and update in the future.

I believe the LSTM is working and plotting. there is a issue with the dates in the date frame that needs to be fixed. once corrected the predictions should be easily appended to the graph and properly shown.